chore: refactor openapi to have less files, more consistency, actually working with API
to do with #65
